code review suggestions
